# api-spec.yml
openapi: 3.1.0
info:
  title: Path AI API
  version: 2.0.0
  description: >
    HTTP API for **Path AI**, a personal productivity system built on Supabase
    Edge Functions (Deno/TypeScript). Each Edge Function maps to a domain and is
    invoked from the React Native client via the Supabase `invoke()` method.


    Authentication is handled automatically by the Supabase client: every request
    carries the anon/publishable `apikey` and a per-user `Authorization: Bearer <JWT>`.
    All data access is additionally protected by PostgreSQL Row Level Security, so a
    user can only ever read or write their own rows.


    Scoring is transactional: routine, goal, and session completions call the
    `calculate_and_award_xp` PostgreSQL function, which applies the streak multiplier
    and multi-virtue bonus and writes append-only rows to the XP ledger.


    **Premium gating** is enforced server-side. Endpoints marked _Premium_ check
    `subscription_tier` before executing and return `402 Payment Required` for
    free-tier users who have exhausted any trial credits.
  contact:
    name: Path AI
  license:
    name: Proprietary — Confidential
servers:
  - url: https://{projectRef}.supabase.co/functions/v1
    description: Supabase Edge Functions base URL
    variables:
      projectRef:
        default: your-project-ref
        description: The Supabase project reference id.
  - url: http://localhost:54321/functions/v1
    description: Local Supabase (supabase start)
security:
  - bearerAuth: []
    apiKeyAuth: []
...
paths:
...
  /goals/backlog:
    get:
      operationId: listBacklog
      tags: [goals]
      summary: Goal backlog items
      responses:
        '200':
          description: Backlog ideas not yet promoted to goals.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BacklogItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: addBacklogItem
      tags: [goals]
      summary: Add to backlog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacklogItemCreate'
      responses:
        '201':
          description: The created backlog item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacklogItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /goals/backlog/{id}:
    patch:
      operationId: updateBacklogItem
      tags: [goals]
      summary: Update backlog item
      description: >
        Edits an unpromoted backlog idea's title, notes, or virtue. Once an item is
        promoted to a goal, edit the resulting goal via `PATCH /goals/{id}` instead;
        editing a promoted item returns `400 Bad Request`.
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacklogItemUpdate'
      responses:
        '200':
          description: The updated backlog item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacklogItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteBacklogItem
      tags: [goals]
      summary: Delete backlog item
      description: >
        Permanently removes a backlog idea. Promoting an item does not delete it;
        use this to dismiss ideas you will not pursue.
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /goals/backlog/{id}/promote:
    post:
      operationId: promoteBacklogItem
      tags: [goals]
      summary: Promote backlog item to active goal
      description: >
        Creates an active goal from the idea and stamps the item with the new goal's
        id; the item's notes become the goal's description and its virtue the goal's
        single tag. Because the goal is created active, the free-tier cap of 3 active
        goals applies and returns `402 Payment Required` when it is reached.
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: Optional overrides applied to the created goal.
              properties:
                is_quest:
                  type: boolean
                  default: false
                due_date:
                  type: string
                  format: date
                base_xp:
                  type: integer
      responses:
        '201':
          description: The newly created goal; the backlog item is marked promoted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Backlog item already promoted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --------------------------------------------------------------- sessions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT for the signed-in user.
    apiKeyAuth:
      type: apiKey
      in: header
      name: apikey
      description: Supabase anon/publishable API key.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    NoContent:
      description: The resource was deleted. No response body.
    BadRequest:
      description: The request body or parameters were invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PaymentRequired:
      description: >
        A premium feature or plan limit was hit. Free users must upgrade (or have
        exhausted trial credits).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The resource does not exist or is not visible to the user (RLS).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    GoalStatus:
      type: string
      enum: [backlog, active, completed, archived]
...
    BacklogSource:
      type: string
      enum: [manual, ai_suggestion, field_log]
...
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 300
        description:
          type: [string, 'null']
        is_quest:
          type: boolean
        status:
          $ref: '#/components/schemas/GoalStatus'
        due_date:
          type: [string, 'null']
          format: date
        base_xp:
          type: integer
        parent_goal_id:
          type: [string, 'null']
          format: uuid
          description: Null for top-level goals; set for subtasks.
        display_order:
          type: integer
        virtue_ids:
          type: array
          items:
            type: string
            format: uuid
        completed_at:
          type: [string, 'null']
          format: date-time
        created_at:
          type: string
          format: date-time
...
    BacklogItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 300
        notes:
          type: [string, 'null']
        source:
          $ref: '#/components/schemas/BacklogSource'
        source_id:
          type: [string, 'null']
          format: uuid
          description: Originating field log or retrospective, if applicable.
        virtue_id:
          type: [string, 'null']
          format: uuid
        promoted_to_goal_id:
          type: [string, 'null']
          format: uuid
          description: Set once promoted to an active goal.
        created_at:
          type: string
          format: date-time
    BacklogItemCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 300
        notes:
          type: string
        source:
          $ref: '#/components/schemas/BacklogSource'
        source_id:
          type: string
          format: uuid
        virtue_id:
          type: string
          format: uuid
    BacklogItemUpdate:
      type: object
      description: Partial update; only supplied fields are changed.
      properties:
        title:
          type: string
          maxLength: 300
        notes:
          type: [string, 'null']
        virtue_id:
          type: [string, 'null']
          format: uuid

    # ---------------------------------------------------------- sessions
...